refactor: 예약 시간 테이블 추가에 따른 테스트 코드 변경
