+++ src/test/java/roomescape/repository/reservation/H2ReservationDaoTest.java
@@ -15,7 +15,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.time.ReservationTime;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 class H2ReservationDaoTest {
@@ -66,21 +67,37 @@ class H2ReservationDaoTest {
     }
 
     Reservation createReservationStuff(String name) {
-        return new Reservation(name, LocalDate.now(), LocalTime.now().withNano(0));
+        return new Reservation(name, LocalDate.now(), createReservationStuff());
     }
 
     void save(final Reservation reservation) {
         JdbcTemplate jdbcTemplate = applicationContext.getBean(JdbcTemplate.class);
-        String sql = "INSERT INTO reservation (name, date, time) VALUES (?, ?, ?)";
+
+        String sql = "INSERT INTO reservation (name, date, time_id) VALUES (?, ?, ?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
             ps.setString(1, reservation.getName());
             ps.setDate(2, Date.valueOf(reservation.getDate()));
-            ps.setTime(3, Time.valueOf(reservation.getTime()));
+            ps.setLong(3, reservation.getTime().getId());
             return ps;
         }, keyHolder);
 
         reservation.setId(keyHolder.getKey().longValue());
+    }
+
+    ReservationTime createReservationStuff() {
+        ReservationTime reservationTime = new ReservationTime(LocalTime.now().withNano(0));
+        JdbcTemplate jdbcTemplate = applicationContext.getBean(JdbcTemplate.class);
+        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setTime(1, Time.valueOf(reservationTime.getStartAt()));
+            return ps;
+        }, keyHolder);
+
+        reservationTime.setId(keyHolder.getKey().longValue());
+        return reservationTime;
     }
 }
